Add example code for create_anzats()
